style: admin and synonym pages
