+++ src/App.jsx
@@ -9,8 +9,8 @@ import SignUp from './pages/SignUp/SignUp';
 import Login from './pages/Login/Login';
 import VerifyOTP from './pages/VerifyOTP/VerifyOTP';
 import Dashboard from './pages/Dashboard/Dashboard';
-const SynonymAdminPanel = lazy(() => import('./pages/SynonymAdminPanel'));
-const AdminPanel = lazy(() => import('./pages/AdminPanel'));
+const SynonymAdminPanel = lazy(() => import('./pages/Admin/SynonymAdminPanel'));
+const AdminPanel = lazy(() => import('./pages/Admin/AdminPanel'));
 import './App.css';
 
 function App() {
