+++ src/pages/Home/Home.css
@@ -140,78 +140,6 @@
   
 
 /* ======================
-     GRADIENT BLOB
-====================== */
-
-.gradient-blob-top {
-    position: absolute;
-    border-radius: 60% 40% 30% 70% / 100% 85% 92% 74%;
-    background: var(--accent-1-dark);
-    background: linear-gradient(90deg, var(--accent-2-dark), var(--accent-1-dark));
-    top: 0%;
-    left: 0%;
-    height: 15rem;
-    width: 15rem;
-    opacity: 1;
-    filter: blur(35px);    
-    z-index: -1;
-    animation: animate 25s cubic-bezier(0.68, -0.55, 0.27, 1.55) infinite alternate forwards;
-    pointer-events: none;
-}
-
-.gradient-blob-bottom {
-    position: absolute;
-    border-radius: 60% 40% 30% 70% / 100% 85% 92% 74%;
-    background: var(--accent-1-dark);
-    background: linear-gradient(270deg, var(--accent-2-dark), var(--accent-1-dark));
-    bottom: 50%;
-    right: 0%;
-    height: 20rem;
-    width: 20rem;
-    opacity: 1;
-    filter: blur(35px);    
-    z-index: -1;
-    animation: animate 25s cubic-bezier(0.68, -0.55, 0.27, 1.55) infinite alternate-reverse forwards;
-    pointer-events: none;
-}
-
-.gradient-blob-footer {
-    position: absolute;
-    border-radius: 60% 40% 30% 70% / 100% 85% 92% 74%;
-    background: var(--accent-1-dark);
-    background: linear-gradient(180deg, var(--accent-2-dark), var(--accent-1-dark));
-    bottom: 0%;
-    left: 15%;
-    height: 20rem;
-    width: 20rem;
-    opacity: 1;
-    filter: blur(35px);    
-    z-index: -1;
-    animation: animate 37.5s cubic-bezier(0.68, -0.55, 0.27, 1.55) infinite alternate-reverse forwards;
-    pointer-events: none;
-}
-
-@keyframes animate {
-    0% {
-      transform: translate(40%, -25%);
-      border-radius: 60% 40% 30% 70% / 100% 85% 92% 74%;
-    }
-    50% {
-      transform: translate(0%, 13%);
-      border-radius: 20% 71% 47% 70% / 81% 15% 22% 54%;
-      rotate: 41deg;
-      scale: 1.15;
-    }
-    100% {
-      transform: translate(-45%, 39%);
-      border-radius: 100% 75% 92% 74% / 60% 80% 30% 70%;
-      rotate: -60deg;
-      scale: 1.05;
-    }
-}
-
-
-/* ======================
      RESPONSIVENESS
 ====================== */
